+++ src/index.ts
@@ -1,21 +1,29 @@
-export {urlCaptureGroup, defaultUserMatcher} from './capture';
+export {
+  urlCaptureGroup,
+  urlRegExp,
+  profileReplacement,
+  defaultUserMatcher,
+} from './capture';
 
 export {allSocialNetworks, defaultSocialNetworks} from './data';
 export {socialNetworkPrefixes} from './prefixes';
-
 export {Socialite} from './socialite';
+
+export {
+  filterNullishValuesFromObject,
+  filterNetworkProperties,
+  getNetworkFromDomain,
+  constructFullLineRegExp,
+  mergeRegExp,
+  getUrlGroups,
+  getUrlWithSubstitutions,
+} from './utilities';
 
 export {MatchUserSource, UrlCaptureId} from './types';
 export type {
   UrlAnatomy,
+  SocialProfile,
   SocialNetwork,
   SocialNetworkProperty,
   SocialNetworkProperties,
-  SocialProfile,
 } from './types';
-
-export {
-  filterNullishValuesFromObject,
-  mergeRegExp,
-  constructFullLineRegExp,
-} from './utilities';
